Update DEVELOPMENT.MD and goplay setup
